+++ RPG Noelf/RPG Noelf/Assets/Scripts/InventoryScripts/Weapon.cs
@@ -17,7 +17,7 @@ namespace RPG_Noelf.Assets.Scripts.InventoryScripts
     {
         public int bonusDamage { get; set; }
         public int bonusDefense { get; set; }
-        public float armorPenetration { get; set; }// tArmor - ArmorPenetration =  new armadura
+        public double armorPenetration { get; set; }// tArmor - ArmorPenetration =  new armadura
         //public Recipe weaponRecipe;
         public TypeWeapon tpWeapon { get; }
         public Weapon(TypeWeapon tWeapon, int goldValue, int amount, string name, bool isStackable, Category categoria, int itemID, string pathImage) :
